rename model/loss to model/loss_and_metric / move metric.py to loss_and_metric
